// src/app/models/user.model.ts
//********************************************************************** */
//********************************************************************** */
//********************************************************************** */

export interface User {
  id: number;
  name: string;
  lastName: string;
  email: string;
  phone: number;
  address: string;
  role: string;
}

//************************************************ */
//************************************************ */

export interface AdminUser extends User {
  userName: string;
  password: string;
}

//************************************************ */
//************************************************ */

export interface FormUser extends Omit<User, 'id'> {}

//********************************************************************** */
//********************************************************************** */
//********************************************************************** */